+++ VisioAutomation_2010/VisioAutomation/Shapes/CustomPropNameCellsPair.cs
@@ -1,0 +1,17 @@
+﻿
+namespace VisioAutomation.Shapes
+{
+
+    internal class CustomPropNameCellsPair
+    {
+        public string Name;
+        public CustomPropertyCells Cells;
+
+        public CustomPropNameCellsPair(string name, CustomPropertyCells cells)
+        {
+            this.Name = name;
+            this.Cells = cells;
+        }
+    }
+
+}
