+++ VisioAutomation_2010/VisioAutomation/VisioAutomation.csproj
@@ -116,6 +116,7 @@
     <Compile Include="ShapeSheet\Writers\WriteRecord.cs" />
     <Compile Include="ShapeSheet\Writers\WriteRecordList.cs" />
     <Compile Include="ShapeSheet\Streams\StreamArray.cs" />
+    <Compile Include="Shapes\CustomPropNameCellsPair.cs" />
     <Compile Include="Shapes\CustomPropertyType.cs" />
     <Compile Include="ShapeSheet\Internal\ArraySegmentReader.cs" />
     <Compile Include="ShapeSheet\Internal\ArraySegment.cs" />
